Bufferline: override default to stop lazy loading
By default the plugin is lazy loaded so will not start until more than 1 buffer is open. This means that when only one buffer is open the name of the file open in the buffer is not shown. Overrided to load the plugin on start up so file name can be seen