Jumplist: map tabufline <TAB> to <C-i> so jump forward works again
for some weird reason <TAB> and <C-i> are indistinguishable so mapping <TAB> stop <C-i> from working and therefore jumping forward in the jump list no longer works. In NvChad core <TAB> is mapped in the tabufline plugin to go switch to the next buffer. Overriding the plugin's <TAB> mapping to <C-i> fixes the jump forward problem and avoids having to make edits to the neovim core.